Move client-compatible series functions into sarc.client: - load_job_series - update_job_series_rgu - compute_cost_and_waste - compute_time_frames
