+++ sarc/client/__init__.py
@@ -1,4 +1,10 @@
 from .job import count_jobs, get_available_clusters, get_job, get_jobs
+from .series import (
+    compute_cost_and_waste,
+    compute_time_frames,
+    load_job_series,
+    update_job_series_rgu,
+)
 from .users.api import get_user, get_users
 
 __all__ = [
@@ -8,4 +14,8 @@ __all__ = [
     "get_jobs",
     "get_user",
     "get_users",
+    "load_job_series",
+    "update_job_series_rgu",
+    "compute_time_frames",
+    "compute_cost_and_waste",
 ]
